Onboarding: block list for java and dotnet
