+++ tests/onboarding/infra_provision/provision_onboarding_host_block_list_dotnet.yml
@@ -1,11 +1,10 @@
 ami: 
-- name: ubuntu-x86-22.04
- # ami_id: ami-0557a15b87f6559cf
-  ami_id: ami-007855ac798b5175e
-  instance_type: t2.medium
-  user: ubuntu
-  os_type: linux
-  os_distro: deb
+  - name: amazon-linux-dotnet
+    ami_id: ami-005b11f8b84489615
+    instance_type: t2.medium
+    user: ec2-user
+    os_type: linux
+    os_distro: rpm
 
 init-config:
   install: !include tests/onboarding/infra_provision/includes/provision_init_vm_config.yml
@@ -31,21 +30,17 @@ installation_checks:
   install: !include tests/onboarding/infra_provision/includes/provision_installation_checks.yml
 
 language-variants:
-  - name: OpenJDK11
-    version: 11
-    install:
-      - os_type: linux
-        os_distro: deb
-        command: sudo apt-get -y update && sudo apt-get -y install openjdk-11-jdk-headless
-
+  - name: dotnet
+    version: 6
+    install: 
       - os_type: linux
         os_distro: rpm
-        command: sudo amazon-linux-extras install java-openjdk11 || sudo dnf -y install java-11-amazon-corretto-devel
+        command: dotnet --info || sudo yum install -y dotnet-sdk-6.0
 
 weblogs:
-  - name: test-app-java
+  - name: test-app-dotnet
     supported-language-versions:
-      - 11
+      - 6
     install:
       - os_type: linux
         command: |
